requirements.txt
Flask==3.0.0
groq==0.1.0
python-dotenv==1.0.0